+++ src/Tada.TemplatePack/templates/service/basic/src/3.Services/TadaSourceName.Services.Tests/TadaTemplateNames/GetTadaTemplateNameServiceTests.cs
@@ -1,4 +1,8 @@
+using TadaSourceName.Domain.Core;
 using TadaSourceName.Domain.Services.TadaTemplateNames.Models;
+using TadaSourceName.Infrastructure.Database.Entities;
+
+using Moq;
 
 namespace TadaSourceName.Services.Tests.TadaTemplateNames;
 
@@ -9,29 +13,50 @@ public class GetTadaTemplateNameServiceTests : TadaTemplateNameServiceFixture
 
     }
 
-    [Theory]
-    [InlineData("00000000-0000-0000-0000-000000000000")]
-    public async Task GetTadaTemplateName_Success(string id)
+    [Fact]
+    public async Task GetTadaTemplateName()
     {
-        var result = await context.GetTadaTemplateName(id);
+        #if(use_repository)
+        var expected = tadatemplatenameFactory.Generate();
+
+        mockTadaTemplateNameRepository.Setup(x => x.GetTadaTemplateName(It.IsAny<Guid>()))
+            .ReturnsAsync((Guid tadatemplatenameId) => expected);
+
+        var result = await context.GetTadaTemplateName(expected.TadaTemplateNameId.ToString());
 
         Assert.NotNull(result);
-
-        //todo: check that result is returning the correct TadaTemplateName
+        Assert.Equal(expected.TadaTemplateNameId.ToString(), result.Id);
+        #else
+        throw new NotImplementedException();
+        #endif
     }
 
     [Fact]
-    public async Task ListTadaTemplateNames_Success()
+    public async Task ListTadaTemplateName_Success()
     {
-        var searchFilter = new ListTadaTemplateNameRequest
-        {
+        #if(use_repository)
+        var expected = tadatemplatenameFactory.Generate(50);
 
-        };
+        mockTadaTemplateNameRepository.Setup(x => x.ListTadaTemplateNames(It.IsAny<BaseListRequest>()))
+            .ReturnsAsync((BaseListRequest request) => new PagedList<TadaTemplateName>(expected, 50, 0, 50));
+
+        var searchFilter = new ListTadaTemplateNameRequest();
 
         var result = await context.ListTadaTemplateNames(searchFilter);
 
         Assert.NotNull(result);
 
-        //todo: check that result is returning the correct TadaTemplateNames
+        expected.ForEach(expectedItem =>
+        {
+            var item = result.First(resultItem => resultItem.Id == expectedItem.TadaTemplateNameId.ToString());
+            Assert.NotNull(item);
+        });
+        Assert.Equal(50, result.Count);
+        Assert.Equal(0, result.Page);
+        Assert.Equal(1, result.Pages);
+        #else
+        throw new NotImplementedException();
+        #endif
     }
+
 }