+++ src/Tada.TemplatePack/templates/service/basic/src/3.Services/TadaSourceName.Services/TadaTemplateNames/Validators/CreateTadaTemplateNameValidator.cs
@@ -1,0 +1,13 @@
+using FluentValidation;
+
+using TadaSourceName.Domain.Services.TadaTemplateNames.Models;
+using TadaSourceName.Infrastructure.Database.Repositories.TadaTemplateNames;
+
+namespace TadaSourceName.Services.TadaTemplateNames.Validators;
+
+public class CreateTadaTemplateNameValidator: AbstractValidator<CreateTadaTemplateNameRequest>
+{
+    public CreateTadaTemplateNameValidator(ITadaTemplateNameRepository tadaTemplateNameRepository)
+    {
+    }
+}