+++ src/Tada.TemplatePack/templates/service/basic/src/3.Services/TadaSourceName.Services/TadaTemplateNames/Validators/UpsertTadaTemplateNameValidator.cs
@@ -1,0 +1,13 @@
+using FluentValidation;
+
+using TadaSourceName.Domain.Services.TadaTemplateNames.Models;
+using TadaSourceName.Infrastructure.Database.Repositories.TadaTemplateNames;
+
+namespace TadaSourceName.Services.TadaTemplateNames.Validators;
+
+public class UpsertTadaTemplateNameValidator: AbstractValidator<(string tadatemplatenameId, UpsertTadaTemplateNameRequest request)>
+{
+    public UpsertTadaTemplateNameValidator(ITadaTemplateNameRepository tadatemplatenameRepository)
+    {
+    }
+}