+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 
 * 1 Add Postgres support
 * 28 Add how to docs
+* 31 service validators